# Merge go1.5+ assembly