--- a/.github/workflows/go.yml
+++ b/.github/workflows/go.yml
@@ -18,6 +18,7 @@ jobs:
           - armv6
           - armv7
           - aarch64
+          - 386
           - x64
         go:
           - '1.5'
@@ -48,6 +49,8 @@ jobs:
           # with newer C compilers. A number of fixes were backported to
           # go1.4-bootstrap but not any released version. See
           # https://github.com/golang/go/compare/go1.4.3...release-branch.go1.4.
+          #
+          # Race builds aren't supported on linux/386.
           #
           # Cross-compilation became possible in go1.5 with the removal of C
           # code from the compiler. See https://go.dev/doc/go1.5#c.
@@ -128,17 +131,46 @@ jobs:
       - name: 'Build with ${{ matrix.go }}'
         if: ${{ matrix.arch == 'x64' }}
         run: go build -v ./...
+      - name: 'Build with ${{ matrix.go }}'
+        if: ${{ matrix.arch == '386' }}
+        env:
+          GOARCH: 386
+        run: go build -v ./...
+      - name: 'Check that Get is inlined with ${{ matrix.go }}'
+        if: ${{ matrix.arch == 'x64' }}
+        run: |
+          if echo -e '${{ matrix.go }}\n1.12' | sort -V | head -n1 | xargs test 1.12 = ; then
+            go build -gcflags='-m' 2>&1 | grep 'can inline Get$' > /dev/null
+          fi
+      - name: 'Check that Get is inlined with ${{ matrix.go }}'
+        if: ${{ matrix.arch == '386' }}
+        env:
+          GOARCH: 386
+        run: |
+          if echo -e '${{ matrix.go }}\n1.12' | sort -V | head -n1 | xargs test 1.12 = ; then
+            go build -gcflags='-m' 2>&1 | grep 'can inline Get$' > /dev/null
+          fi
       - name: 'BuildRace with ${{ matrix.go }}'
         if: ${{ matrix.arch == 'x64' && !matrix.x64_race_broken }}
         run: go build -race -v ./...
       - name: 'Test with ${{ matrix.go }}'
         if: ${{ matrix.arch == 'x64' }}
         run: go test -v ./...
+      - name: 'Test with ${{ matrix.go }}'
+        if: ${{ matrix.arch == '386' }}
+        env:
+          GOARCH: 386
+        run: go test -v ./...
       - name: 'TestRace with ${{ matrix.go }}'
         if: ${{ matrix.arch == 'x64' && !matrix.x64_race_broken }}
         run: go test -race -v ./...
       - name: 'Bench with ${{ matrix.go }}'
         if: ${{ matrix.arch == 'x64' }}
+        run: go test -bench=. -benchmem -v ./...
+      - name: 'Bench with ${{ matrix.go }}'
+        if: ${{ matrix.arch == '386' }}
+        env:
+          GOARCH: 386
         run: go test -bench=. -benchmem -v ./...
       - name: 'BenchRace with ${{ matrix.go }}'
         if: ${{ matrix.arch == 'x64' && !matrix.x64_race_broken }}
@@ -177,7 +209,7 @@ jobs:
         if: ${{ matrix.arch == 'aarch64' && !matrix.arm64_race_unsupported && matrix.arm64_unsupported_vma_range }}
         run: rm goid.race.test
       - name: 'Test and Bench with ${{ matrix.go }} on ${{ matrix.arch }}'
-        if: ${{ matrix.arch != 'x64' && !matrix.qemu_emulation_broken }}
+        if: ${{ matrix.arch != '386' && matrix.arch != 'x64' && !matrix.qemu_emulation_broken }}
         uses: uraimo/run-on-arch-action@v2
         with:
           arch: ${{ matrix.arch }}