… goid_go1.5_amd64.go
… goid_go1.5_arm.s
… goid_go1.5_arm64.s