@tamird 2023-05-11
Added CI coverage for inlining as well.